+++ src/components/common/tabs/Tabs.vue
@@ -1,0 +1,18 @@
+<template>
+  <a-tab-pane :key="key" :tab="tab"></a-tab-pane>
+</template>
+
+<script>
+export default {
+  props: {
+    key: {
+      type: String,
+      required: true
+    },
+    tab: {
+      type: String,
+      required: true
+    },
+  },
+}
+</script>
