+++ src/components/communications/top/StartCall.vue
@@ -1,0 +1,7 @@
+<template>
+  <router-link to="video-call">
+    <a-button class="btn blueBtn">
+      {{ $t('communications.startCall') }}
+    </a-button>
+  </router-link>
+</template>
